Improve the 'permissions' test
`sendPermissions` expects `this` to be a select element with an id and a
value. Relying on `this` is not a great idea. In the test, `this` is
actually `window` so the test setup needed to set an id and value on the
window object. So, to avoid using `this` we can actually use the target
of `event` but the test was also setting up the even with a bad target
(a string instead of an actual element) so that needed to be fixed too.
Also the fixture html was a bit weird and probably invalid.