Make Invalid Token warnings more user friendly (#8445)
* The session token validation implementation currently gives an empty white screen with "Invalid token" text. This is total uninformative to a normal user.
This update implicitly redirects to the previous page if the token is not valid. To prevent implicit redirect pass `false` as the second argument to `JSession::validateToken()`

To make use of this method we need to replace any occurrence of `JSession::checkToken('post') or jexit(JText::_('JINVALID_TOKEN'));` with `JSession::validateToken('post');`
Example is set up in com_users > login method.

* Moved the checkToken method to JControllerLegacy as it is the more closer context to the usage.
This allows implicit redirect to the REFERER page in case of token mismatch. In all other case it is same as calling JSession::checkToken() directly.

Changes made in the frontend components to make this change test-able.

Note: We intend to discourage the `die` usage everywhere which leads the the said uninformative white page.

* Fix wrong function call

* use automatic since tag macro

Thanks @andrepereiradasilva