JApplicationWeb::setHeader/sendHeaders behavior (#9836)
* Update web.php

setHeader $replace previously did nothing other than delete a value - even when false it allowed a duplicate to be created which effectively replaces the original.

This modification searches for an existing header with the same name and only inserts it if it doesn't already exist, or if $replace is true.

* JApplicationWeb::setHeader behavior

Fix Travis CI error regarding invalid foreach

* JApplicationWeb::setHeader behavior

Formatting changes

* JApplicationWeb::setHeader behavior

Formatting issues

* Fix testSetHeader



in line 1314 "$this->class->setHeader('foo', 'car');", the method is called without the $replace argument, meaning it uses the default of "false".

The way headers are set, allowing a duplicate here means the header is replaced in JApplicationWeb::header.  If $replace isn't enforced in setHeader then duplicates mean the header gets replaced anyway.

The attempted entry is removed from the assertEquals test.

* JApplicationWeb -setHeader and sendHeader upgrade

Current Joomla header operation causes a duplicate header to replace an original header even when $replace = false.  The replacement is a result of interaction between setHeader and sendHeader.

setHeader happily adds duplicate entries when $replace = false
sendHeader happily replaces headers because its default is to replace

This commit alters both:

setHeader now checks an internal list of headers which may only have one value.
If the header is a single value only, the header array is updated only if $replace = true and the update is ignored otherwise
If the header is not a single value, the header is replaced if $replace = true, otherwise it is added as a multiple value 

sendHeader now creates a new array of values, aggregating values under the header name.  When the header is sent to the header() method, its value is imploded on ', '

* Returning test to original state

this should be removed from the branch but I'm not sure how to do it.

* JApplicationWeb::setHeader behavior

Update to fix status header

* JApplicationWeb -setHeader and sendHeader upgrade

Fixing for test output

* JApplicationWeb -setHeader and sendHeader upgrade

formatting

* JApplicationWeb -setHeader and sendHeader upgrade

formatting

* JApplicationWeb -setHeader and sendHeader upgrade

formatting

* JApplicationWeb -setHeader and sendHeader upgrade

refinement - removed unnecessary parens

* JApplicationWeb -setHeader and sendHeader upgrade

performance enhancement

* JApplicationWeb -setHeader and sendHeader upgrade

performance enhancements, 2 fewer foreach iterations for the headers

* JApplicationWeb -setHeader and sendHeader upgrade

formatting fix

* JApplicationWeb -setHeader and sendHeader upgrade

Missed a variable change i sendHeader

* JApplicationWeb -setHeader and sendHeader upgrade

formatting fix

* Update web.php

* JApplicationWeb -setHeader and sendHeader upgrade

fix formatting

* JApplicationWeb -setHeader and sendHeader upgrade

Adding ability to unset (remove) an existing header using $replace=true and $value=null

setHeader('name', null, true) now removes the header entirely, where previous behavior would have sent the header with an empty value.

* JApplicationWeb -setHeader and sendHeader upgrade

formatting fix