[Easy fix] Typo - a trademark should not be altered
Easy fix on review.
As a trademark, `The Joomla! Project`, even in core language strings should not be altered (no exceptions as a legal matter).
I know no one cares (and TBH me neither) but you can't ask others to respect a rule if we don't.

Previous:
`The Joomla project`
New:
`The Joomla! Project`