com_finder outputs empty p tag with no description
### Steps to reproduce the issue
Set description on article or anything other item to nothing. In my scenario it was using Docman where they are just documents with no content.

### Expected result
html not to be rendered if there isn't a description to render

### Actual result
Renders a <p> html tag with no content, causes excess padding and just a mess on most sites.


### System information (as much as possible)
J3.6.5 / J3.7.0Alpha2.

### Additional comments

this is just a simple change to make the output of search results which have no description inside them to not render excess html.