[com_fields] Add option to show or hide field label (#13752)
* Added option to hide field label

* Added variable and if statement to hide label

* Added language strings for the Hide Label option

* Updated to change the if statement as per comment

* missed a semicolon

* changing switch to Show/Hide

* Changing switch to Show/Hide

* Changing switch to Show/Hide

* missed a language string change

* Changed if statement

Since the default is Show, if it's not saved in the DB, the label won't show. Changed the if from ==1 to !=0 so it shows without any entry in the params in the db.

* Reversed the if statement

I tested on CURRENT fields (not new ones) but I had changed the params name, so it didn't stick. All my fault :)