Made reference to fulltext alt instead of introtext alt
It seems that this makes reference to the image_fulltext_alt but the rest of the code seems to be targeting the intro text.  Shoot me down if I'm wrong, I only noticed this because I was trying to merge a conflict.

I would also question (and I really don't know the pros/cons) as to whether this could use a JLayout instead.  The pro in my eyes would be code maintainability and standardisation.  There may be an argument against this I really am not skilled enough to know.