Admin menu Systemm information
When creating a new Admin menu item the avaiable menu type are displayed in component alphabetical order.

Currently the System Information component will be displayed at the very top of the list as it is named Administrator-System Information. There is no need for this to prefixed with Administrator and by removing the prefix it is moved down the list into a more logical position