Media Manager remove Root text
In media manager the path is prefixed wth the word Root: Is that needed? It doesnt provide any information and is confusing.

This came up in a training session where the user was looking for a folder called root

This PR simply removes the word Root: as suggested by commenters here #14597