Remove Hathor template from Joomla!
The Hathor administrator template overrides are not maintained since long and as a consequence it now appears completely broken in J3.7.
Attempts should have been made to fix the issues but its a long list to do. We are not even sure if even someone is still using it.

Therefore, I propose to remove this from Joomla as it is not usable anyway.

However, if we can still fix it I am all in to my part of contribution and help it stay here until we have 4.0 at least.