[4.0] Install test data
This is a TEMPORARY pr while we are in early development stage and only have one sample data set available for testing.

It changes the text on the screen from
#### Install Sample Data

to
#### Test English (GB) Sample Data

This is to avoid the situation where we have had bug reports about the "default install" when there is not default sample data to be installed prepared yet. Remember the testng sample data is never shipped wwith Joomla

(Yes i am aware that the testing data needs updating but thats another issue)