Offline page not working in the Beez3 template (#15711)
* Offline page not working on Beez3 template

The Beez3 template doesn't have an error.php so the default system template error.php is loaded. In that case the paths to the CSS files are incorrect as they are looking for the CSS files in the Beez3 template CSS folder and not in the system CSS folder.

* Removed obsolete code

Removed 'relative' => false as false is the default value.