[a11y] JHtml grid improvements (Work in progress)
This is a work in progress I could really do with some assistence to complete this - see below

### Issue
In all list views we start the row with a select box. This select box has no label which is not good AND when using assistive technology such as a screen reader it will only indicate the word "select". It doesnt indicate what you are selecting. As the title (the item you are selecting) is several steps away from the select box you are stuck with a select with no meaning.

### Fixing the issue
 [x] add a label for the input
 [x] add hidden text for the label (screenreader only with the class element invisible)
 [ ] add the title of the item in the label - currently I only added the item id.

### With ths PR
There is now a label with the input and it says "select 3" where 3 is the item id. This should be the title.

### help required
 [ ] use the item title not the id in the label
 [ ] update the doc blocks
 [ ] update the unit tests

reference for the a11y technique used http://www.coolfields.co.uk/2011/04/accessible-forms-should-every-input-have-a-label/