If article page does not exist, throw 404 (#11217)
This is in addition to https://github.com/joomla/joomla-cms/pull/5306 which was merged some time ago. While that pull request fixed issue with invalid article pages when article still had some pages, there was still a problem if you would completely remove page breaks from an article, because then the exception would never be thrown, it was too low in the code. Now if plugin detects that there are no pagebreaks in the article body, it will also additionaly check if there is a request for a subpage, and if yes throw 404.

To test, try opening some article that does not have page breaks and add to the end of the url ?start=xx where xx is any number. You should see the article normally. Apply the patch and make sure the same link gives 404 page.