Remove redundant word (#14607)
Pull Request for #10971

As requested by @ot2sen and others this PR removes the redundant word successfully from success messages. It's just not needed at all. You can't have a success message that says you "unsuccessfully" did something so there is no need to use the word.

This can be merged after 3.7.0 if requested - this is upto @bakual as commented at #10971