Move core extensions list to new helper class
Move the static list of core extensions used in script.php for updating
the manifest caches to a new helper class for making it possible to be
used by other functions, too, and add a function for check if an
extension is core or not.
This provides a safe way to detect core extensions, while checking if
extension ID is greater than a certain value is not a safe way.