Replacement of Austrian Flag (#16800)
* Deletion of Austrian Flag to be Replaced

The current Austrian flag (de_at.gif) is a Service/State flag used in the Government buildings and officials.  The Civilian flag does not the insignia in the middle.  So, a new flag image has to be used in the place of the current Flag.

* Revised Austrian Flag image (de_at.gif)

Revised Austrian Flag image (de_at.gif) - This is a Civilian Flag instead of Service Flag.