Missing word in string
### Before

JGLOBAL_SHOW_CREATE_DATE_DESC="If set to Show, the date and time an Article was created will be displayed. This a global setting but can be changed at Menu and Article levels."

### After
JGLOBAL_SHOW_CREATE_DATE_DESC="If set to Show, the date and time an Article was created will be displayed. This is a global setting but can be changed at Menu and Article levels."