[4.0] com_banners remove popover
This is a proof of concept for #17044 to remove the redundant popovers.

PURELY for the purpose of testing this concept I have quickly displayed the remaining popover as inline text - THIS PART IS NOT FOR REVIEW - it is just to demonstrate how few remaining descriptions we will have

In addition you will notice that the remainining descriptions are not very good. They are manuals not tips, hints or descritpions and can be drastically reduced or rewritten. AGAIN this text is not for review here