Banned text strings consistency (#17108)
The text used for the banned fields in com_contacts is different in the commponent options and the individual contact options. Irrelevant of if the information is needed or not we should be consistent. So on the basis that it is important to have the extra information I have updated the individual contact strings to match the global ones.