[4.0] Change vote option to switcher
This PR changes the vote setting in the article options to the new switcher class for options

An added benefit it that before when the plugin was not enabled the select list was empty. Now if the plugin is not enabled there is nothing displayed in the values. NOTE if someone knows how to improve this perhaps by hiding the label as well that would be great