Allow disabling the Joomla! localhost database security check. (#17298)
As Joomla! provides a Docker image, the chance of using "localhost" is never. Therefore if the container is scheduled in an environment where you cannot access the container (not that you should ever need to) then you need to be able to bypass this check. This environment variable can be placed in the containers build file and will disable checks meaning you can install hassle free.

This check still keeps the installation secure as you'd need the same access to the server to delete the file.