Required fields cleanup
A form should have as few required fields as possible. In many cases we have marked a field as required even though it can never be empty because it has a default value so we dont need to mark it as required.

In addition we have the field banner clients email marked as required but there is no functional reason for this to be required