Update placeholder with module name from manifest.
During a module install, when copying base files fails and exception is
thrown with a language definition. This language definition contains a
placeholder but is no replaced.

This commit will now make it replace the placeholder with the name from the
manifest.

The language definition is "JLIB_INSTALLER_ABORT_MOD_COPY_FILES"
The text at this time is "Module %s: Could not copy files from the
source."