Fix PHP warning if an old type menus exist
I just noticed that error log on one of my sites quickly grew to over 100 MB. It was always the same error:

_Trying to get property of non-object in /*******/public_html/libraries/cms/component/router/rules/menu.php on line 179._

I was not the only one with this problem, here is a related post on the forum: https://forum.joomla.org/viewtopic.php?t=949743

It would happen for example if when migrating from an old version you had a menu of type "Content Section", and it wouldn't matter that the menu isn't displayed anywhere on the new version. It may happen under some other circumstances as well. The change I made should keep the condition of the if statement unchanged, except that it won't throw an error if the variable is uninitialized.