Load component language files from both system-wide and applications-specific folders
PR for gh-17272

Pinging @b2z @laoneo @infograf768 and @Bakual

# Summary of changes

Joomla! currently loads language files from the system-wide folder (`language` or `administrator/language`). If no files are found there then, and only then, will Joomla! look in the component folder (`components/com_example/language` or `administrator/components/com_example/language`).

The problem is that a component can only ship its language files inside the component directory, including the default language (en-GB). If the user now installs a language pack for this extension they will get system-wide translation files installed, let's say for fr-FR. If that translation (in our case fr-FR) is incomplete the user will receive untranslated strings (e.g. COM_EXAMPLE_SOMEVIEW_LBL_WHATEVER) instead of the default (en-GB) translation. This happens because of Boolean short-circuit evaluation which causes Joomla! to stop looking for language files once the system-wide files are loaded.

This PR removes the Boolean `OR` and replaces it with two separate lines. The order of the lines, as discussed in gh-17272, is such that system-wide language files override component-specific language files.

# Documentation changes required

You _may_ want to document that default language files can be placed in either system-wide or component-specific folders starting Joomla! 3.8.0 (or whatever version this is merged) whereas in previous versions they HAD TO be placed in the system-wide folders. Just so that this bit of institutional knowledge is not lost to the sands of time and the interwebz ;)