Complete the removal of weblinks from stats modules
In #13389 the code for weblinks in the stats module was rempved but the description remains in the module. This is confusing and so this PR removess the text about weblinks from the description. Its not needed just as there is no text in the description about having plugins installedd that will inject other stats into that module