[a11y] Readmore links
Alternative to #17469

There are many possible ways to display a readmore link depending on the options chosen
1. Read more
2. Read more with title
3. Read more with title limited to a set number of characters
4. Register to Read more with title (Show unathorised links)
5. Register to Read more with title limited to a set number of characters (Show unathorised links)
6. As above with alternative Read More Text

Option 1 fails a11y as you end up with multiple links on the page all with the same title
Option 3 may fail a11y as the title may be truncated (even inside a word) and this can lose meaning and sometimes result on multiple links with the same title
Option 4 fails a11y as you end up with multiple links on the page all with the same title

To resolve this we can use the aria-label attribute on the link to ensure that assistive technology will always see a unique and meaningful link.

This alternative PR addresses the issues raised by @zwiastunsw
A) does not say that it is a continuation (Read more).
B) when access is for registered users - incorrectly states that this link and specifies where instead of message: Login to see