[4.0] com_redirect remove archive state
In com_redirect we have all the states (published, unpublished, deleted and archived) but just like other components such as com_modules we have no need for the archived state.

Note it was suggested by @mbabker in #17417 that this was a small b/c break so could not go into J3.x

Not sure if I should be doing another PR to the J3 branch to add deprecate notices and if so where?