[com_fields] Automatic Display of Custom Fields. (#17607)
Pull Request for #15204

Original Report
We have the options "After Title", "Before Display","After Display" and "No". Translating this into Dutch I do think it is better to use "Not" instead of "No". I am not native English, but in Dutch it is better. Because of the use of JNO as string I don't have the option to change it in the Dutch translation.

This PR changes it from No to Do not automatically display and it is also a new string so that the dutch can translate it correctly