Made calls consistent
Removed bottom check for featured and copied first check.  Previously it checked for empty but this was an unnecessary check as featured is either on or off.  I think this may have come from earlier discussion as to which method loads quickest in PHP.  I'm going to leave it this way round but if someone thinks it can load quicker then I am happy to change again.