Check position of base uri rather than just existence
The if condition has been updated to check for the position of the base uri in the srcset url, rather than just for the existence of the base uri in the string. This is because it is possible that the base uri string could be repeated somewhere in the url, so the check is only done to see if the base uri is at the beginning of the srcset url.