Check for JUri::base in each srcset url before converting (#17978)
* Check for JUri::base in each srcset url before converting

With reference to - https://github.com/joomla/joomla-cms/pull/15300#issuecomment-329561846, this PR updates the srcset URL conversion to check for the base url before performing the replacement.

## Summary of Changes

As pointed out by @wronax - https://github.com/joomla/joomla-cms/pull/15300#issuecomment-329559714 - if the srcset URL already contained the site base url, then the conversion would add the base url again. This change checks for the base url on each of the srcset urls before performing the replacement, if necessary.

## Testing Instructions

As before, create a new article using the following HTML code, or similar. You will need 3 images of different resolutions, or just three different images of any resolution.

```html
<img src="images/image1.jpg" srcset="images/image3.jpg 2x,images/image2.jpg 1.5x,images/image1.jpg 1x" alt="" />
```

## Expected result

The image displays correctly in the browser as all the srcset values are converted.

## Actual result

With the current sef.php the images will not display correctly in a modern browser. You can check the code using the browser console to see that the srcset values have not all been converted.

## Documentation Changes Required

None

* Fix if condition syntax and update regex

The regex is updated to allow each srcset url to contain spaces in the url, and to not contain a device pixel ratio.

* Check position of base uri rather than just existence

The if condition has been updated to check for the position of the base uri in the srcset url, rather than just for the existence of the base uri in the string. This is because it is possible that the base uri string could be repeated somewhere in the url, so the check is only done to see if the base uri is at the beginning of the srcset url.

* Remove base uri check and update regex

* Update regex again to allow for spaces in srcset url

* Trim whitespace from each srcset url