php 7.2 compatibility mcrypt checks
PR for 18007

We were checking and reporting on the presence of mcrypt. We no longer need to do that as it is no longer a requirement and mcrypt itself is not present in php 7.2 so the checks would fail anyway then.

Didnt see the point in wrapping it in a conditional php version check