Search with a / in the title
PR for #18220

Code by @hikashop-nicholas

### Steps to reproduce the issue

Create a new article with the title "SHIM/SET"
Search for it on your frontend via the Joomla search.

### Expected result

The result page should display a search result for it with a link to the article.

### Actual result

The result page has a search result but without any link.

> Warning: preg_replace(): Unknown modifier 't' in C:\Apache24\htdocs\cms\components\com_search\views\search\view.html.php on line 268

### Test
Apply this PR and the title is displayed