[3.9] remove mcrypt check
We were checking and reporting on the presence of mcrypt. We no longer need to do that as it is no longer a requirement and mcrypt itself is not present in php 7.2 so the checks would fail anyway then.

Didnt see the point in wrapping it in a conditional php version check as we dont need the info any more

See #18169 for the other mcrypt removal