Label element without control
Whenever possible, use the label element to explicitly associate text with form elements. The for attribute of the label must exactly match the id of the form control.

A <label> can be associated with a control either by placing the control element inside the <label> element, or by using the for attribute.

This PR fixes the missing "for" element for the remember me label on com_login

To test go to the login component (not module) and check that the markup for the word remember me is

`<label for "remember">` where remember is the id of the checkbox.

The remember me plugin must be enabled