Help screen invalid markup
Lists must be marked up correctly, meaning they must not contain content elements other than <li> elements. The divider between the items in the list of help pages is an hr which is invalid.

This PR changes that to a li with a class of divider just as in the drop down menus. There is no visual change

Screen readers have a specific way of announcing lists. This feature makes lists clearer to understand, but will only work if lists are properly structured. Properly-structured lists only include <li> content elements.

This resolves an a11y level (A): MUST fix
WCAG Success Criteria:1.3.1 Info and Relationships