Flags alt tags (#18467)
Currently the alt tag for the flags shown (if enabled) for a language is the same as the text next to it. This is not correct as the alt text should not be the same as the text. In addition if an image is only used for decorative purposes then the alt tag should be blank so that screen readers will not announce it.