Update dispatcher.php
This small change allows media fields to be used in user profiles, where a user is not authorized for core.create in com_users - they can be authorized in com_media.  In conjunction with a filesystem plugin, a user can be granted access to com_media a limited area of the filesystem.  This change should not impact anything other than front-end media modals.