Grammar fix
Small fix to correct the grammar/typo as it should be field not fields

### Before
For increased security the edit permission for this SQL fields was set to denied for all non Super Users.

### After
For increased security the edit permission for this SQL field was set to denied for all non Super Users.