Profile plugin fields can be left empty when required on admin edit
## Steps to reproduce the issue
Enable profile plugin. set fields to required for site and admin edit. We used Phone, Company, and website in our test.

## Expected result
Profile plugin fields are required during admin edit.

## Actual result
Can save profile during admin edit when profile plugin fields are empty.

After this PR the fields set to required are marked with a * and have the class required

PR for #14805