Profile plugin fields can be left empty when required on admin edit (#19141)
* Profile plugin fields can be left empty when required on admin edit

## Steps to reproduce the issue
Enable profile plugin. set fields to required for site and admin edit. We used Phone, Company, and website in our test.

## Expected result
Profile plugin fields are required during admin edit.

## Actual result
Can save profile during admin edit when profile plugin fields are empty.

After this PR the fields set to required are marked with a * and have the class required

PR for #14805

* tab

* oops

* oops (reverted from commit 3ec6bd05cf9e51acbdc9f7165cf3c692c4ecd908)