[4.0] Frontend Article submit in mono language
Pr for #19260

I didnt realise that the frontend didnt use the global edit layout so the code to set language to all was not present

## Steps to reproduce the issue
Install J4 with Blog Sample Data
Add a Menu Item Type: Create Article to the Main Menu Blog Menu with Access of Registered.
Select Blog as the Category in the Options Tab for that Menu Item.
Go to the frontend and login (as any UG. I did it as a Super User.)
Click the new Menu Item you just created and add a Title and a bit of text if desired.
Save it.

## Expected result
That the article would save.

## Actual result
Error: Save failed with the following error: Field 'language' doesn't have a default value and the article will not save.