Select not click
Update to en-gb language files to take into account mobile/tablet use where you can not click.

This was first done in #7430 but some have slipped back in

Basically it replaces click with select - Click refers to a mouse and is not relevant if using a tablet or a phone without a keyboard.

NOTES
I have kept banner impressions to refer to clicks
I have kept weblink impressions to refer to clicks

I have kept joomla-update as "one-click" as that was the marketing

Please remember the most important object of the exercise is to make the language files consistent