Show filename at runtime exception
Summary of Changes

if a extension use field type "subform" and the xml file is in the wrong directory, the displayed information contains now the used path to the xml file.

Testing Instructions

use any extension from JED who use field type "subform" and rename/remove the subform xml file.
i have seen this problem, as i developed my own addon on my test system.
But my extension is not online now and also do not have this problem anymore :)

Expected result

if this problem exists, i will see the path to the file additional if j! is in JDEBUG mode

Actual result

i only see, that the file cannot be found

Documentation Changes Required

i think here is no change to do but this should decide the pros of joomla :)