The fix for Joomla progressive cache causing multiple duplicate javascript declarations
The issue has been described and discussed at 
https://github.com/joomla/joomla-cms/issues/5933
https://github.com/joomla/joomla-cms/pull/2558
https://github.com/joomla/joomla-platform/issues/673

In my case I had a problem with JCE mediabox plugin failing once a page has been cached, due to its init script getting duplicated up to 8 times. 

This PR fixes the bugs in two lines that were meant to prevent duplication but the stristr() checks string inclusion in the wrong way: $sdata/$stdata here is the cached script declarations string so it has to be a haystack and be the first argument.