Improvements to Categories List View
The changes for this commit improve the Categories List view by adding
in the capability for limited user groups to only see the Categories
that they have access to and also allows the ability to filter by one or
more of those categories the user can access.

This improves the usability for Joomla for larger sites where we want to
have user groups that can only access certain categories on the site
because currently the default behavior is to display all categories
regardless of whether or not the user has access to them and only
"helps" by graying out the ones the user does not have access to which
really isn't very helpful when you have many categories.

The ability to filter by category is also helpful for a Super User who
might have to assist with various sections of the website. It basically
provides a quick filter to see just the categories that belong to a
certain section of the website to more easily assist with edits that are
requested.