Improvements for Articles List View
The improvements here are similar to the ones made for the Categories
List View.

Basically, for limited users that only have access to certain
categories, the Articles List View currently shows those users all of
the Articles on the website and only "helps" by graying out the ones
they don't have access to, which may not be that helpful if you have a
lot of articles.

Instead, with these additions, those users will now only see the
Articles that are in Categories they have access to and helps make the
backend interface more useful/usable for this type of user.