Toolbar Adjustments for Limited Users
For those users with limited access to certain Categories, unfortunately
the default behavior of the toolbars hides the sometimes useful buttons
for Edit / Publish / Unpublish / Archive / Trash due to the core.edit,
core.edit.own, and core.edit.state permissions not being present for
these users at the Component Level (they are present though in the
individual Categories the user has been assigned however).

Since there wasn't an existing way of having that check "just work" with
the existing access checks, a compromise here would be to detect the
presence of there being items in the list. If so, then go ahead and
display these extra buttons.

With the changes made in the previous two commits, users that do not
have the rights to access the extension likely wouldn't be able to get
into the component's admin area in the first place, and if they can then
they'll likely have some sort of access rights to make edits and changes
to the Categories (or Articles) that they have access to so I don't
believe the changes made in this commit will negatively affect existing
Joomla users.