Categories List View Bugfix When Clearing Filters
This is an issue I noticed during testing of the other commits.

It only happens with the Categories List View since it is designed to
work with multiple extensions.

Currently, if you are in any extension's Categories List View and hit
the "Clear" button it'll redirect you to a URL that doesn't have the
extension parameter in the URL like this:
index.php?option=com_categories&view=categories

The issue isn't immediately noticeable because clicking the "Clear"
button submits the form on the page and this includes a hidden variable
containing the extension name. However, if you go to your Browser's
address bar and click in there and press Enter to reload the page, if
you are a Manager (or one of the users with limited user access) you'll
receive an error. The Administrator and Super User accounts bypass the
check that occurs in the extension entry file so you can't see the issue
with these types of users.

After making this minor tweak to include the extension back into the
form's action attribute, the extension is present in the URL so users
can go to the address and reload without trouble.